feat(ucc-activity-countdown-timer): Renamed the icon file
