+++ plugin-flex-ts-template-v2/src/feature-library/ucc-activity-timer/custom-components/ActivityTimerWithPiP/PIPIcon/ActivityTimerIcons.tsx
